image_handler/pyproject.toml
[tool.poetry]
name = "image-handler"
version = "0.1.0"
description = ""
authors = ["Максим Жданов <exodus.yarwood@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
redis = {extras = ["hiredis"], version = "^5.0.8"}
pillow = "^10.4.0"
pydantic-settings = "^2.4.0"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
